refactor(R2.2): AI 域 1182 行摘出 ReadBookViewModel，状态独立成 ReadAiUiState
ReadBookViewModel 6671 → 5450 行。AI 域（章节摘要 / 划词净化 / 划词重写 /
重写预设）搬进 ReadAiDelegate，自持 MutableStateFlow<ReadAiUiState>，
不再挂在 ReadBookUiState 上——AI 流式输出每秒刷新多次，挂在阅读态上会让
整个 ReadBookUiState 反复 copy。

activeSheet 仍是 ReadBookUiState 的单一持有者：AI sheet 与阅读 sheet 的
互斥、返回键的 `activeSheet != null` 都零改动。delegate 通过 7 个方法的
Host 接口反向触达 sheet 开合 / 关菜单 / 正文重载 / Toast / 章节读取。

逻辑逐行未改，只有机械替换：viewModelScope→scope、6 处外部耦合换 Host
回调、4 个原本内联在 onIntent 里的 preset setter 提成方法。

章节读取（3 处 appDb.bookChapterDao）刻意不跟着搬：
legacyDaoInjectionBaseline 只统计文件名含 ViewModel 的文件，搬进 delegate
会掉进宽松的 legacyUiDaoAccessBaseline，等于把 VM 棘轮上的债洗白。改由
Host 回调，留给 R2.1 随 ReaderSession 一起溶解。两次 getChapter 合并成
一个 Host 方法，基线 36 → 35（棘轮强制下调）。

新护栏 ReadAiDomainBoundaryTest 三条，均已注入回归反证会变红：
1. AI 子状态挂回 ReadBookUiState → 红
2. AI 状态类型回流 ReadBookViewModel → 红
3. ReadAiDelegate 自带 DAO 直连 → 红（gradle verifyConfigArchitecture 也独立拦住）

验证：compileAppDebugKotlin / verifyConfigArchitecture / 417 单测全绿 /
lintAppDebug 0 error / assembleAppDebug 通过。
